Rename packages
Moves the packages from resourcecontainer to rcmediadownloader.

This library is not part of the resource container library and should not collide with the namespace.
