+++ cliapp/src/main/kotlin/org/wycliffeassociates/rcmediadownloader/cli/Main.kt
@@ -1,3 +1,3 @@
-package org.wycliffeassociates.resourcecontainer.media.cli
+package org.wycliffeassociates.rcmediadownloader.cli
 
 fun main(args: Array<String>) = CIApplication().main(args)
