+++ lib/src/main/kotlin/org/wycliffeassociates/rcmediadownloader/data/MediaDivision.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.resourcecontainer.media.data
+package org.wycliffeassociates.rcmediadownloader.data
 
 enum class MediaDivision {
     BOOK,
