+++ lib/src/main/kotlin/org/wycliffeassociates/rcmediadownloader/data/MediaUrlParameter.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.resourcecontainer.media.data
+package org.wycliffeassociates.rcmediadownloader.data
 
 data class MediaUrlParameter(
     val projectId: String,
