+++ lib/src/main/kotlin/org/wycliffeassociates/rcmediadownloader/io/IDownloadClient.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.resourcecontainer.media.io
+package org.wycliffeassociates.rcmediadownloader.io
 
 import java.io.File
